Improvements to deletion logic. Only pushed it through for weight matrix and neuron array. It remains to be implemented for other objects types

+++ src/main/java/org/simbrain/network/groups/Subnetwork.java
@@ -20,6 +20,8 @@ import org.simbrain.network.core.NetworkModelList;
 import org.simbrain.network.core.Neuron;
 import org.simbrain.network.core.Synapse;
 import org.simbrain.network.events.LocationEvents;
+import org.simbrain.network.matrix.NeuronArray;
+import org.simbrain.network.matrix.WeightMatrix;
 import org.simbrain.util.propertyeditor.EditableObject;
 import org.simbrain.workspace.AttributeContainer;
 
@@ -79,6 +81,11 @@ public abstract class Subnetwork extends LocatableModel implements EditableObjec
         modelList.add(model);
         model.setId(getParentNetwork().getIdManager().getAndIncrementId(model.getClass()));
         model.getEvents().onDeleted(m -> {
+            // TODO: Conditional is temporary to test new deletion logic
+            if ((m instanceof WeightMatrix) ||
+                    (m instanceof NeuronArray)){
+                deleteModel(m);
+            }
             if (modelList.getSize() == 0) {
                 delete();
             }
@@ -93,10 +100,14 @@ public abstract class Subnetwork extends LocatableModel implements EditableObjec
         events.fireDeleted();
     }
 
-    public void deleteModel(NetworkModel toDelete) {
+    private void deleteModel(NetworkModel toDelete) {
         modelList.remove(toDelete);
-        toDelete.delete();
-        parentNetwork.getEvents().fireModelRemoved(toDelete);
+        // TODO: Conditional is temporary to test new deletion logic
+        if (!((toDelete instanceof WeightMatrix)
+                || (toDelete instanceof NeuronArray))) {
+            toDelete.delete();
+            parentNetwork.getEvents().fireModelRemoved(toDelete);
+        }
     }
 
     public NetworkModelList getModelList() {
